+++ Develappers.BillomatNet/InvoiceService.cs
@@ -14,6 +14,7 @@ using Invoice = Develappers.BillomatNet.Types.Invoice;
 using InvoiceDocument = Develappers.BillomatNet.Types.InvoiceDocument;
 using InvoiceItem = Develappers.BillomatNet.Types.InvoiceItem;
 using InvoiceMail = Develappers.BillomatNet.Types.InvoiceMail;
+using InvoiceComment = Develappers.BillomatNet.Types.InvoiceComment;
 
 namespace Develappers.BillomatNet
 {
@@ -385,6 +386,50 @@ namespace Develappers.BillomatNet
                 InvoiceMail = model.ToApi()
             };
             return PostAsync($"/api/invoices/{id}/email", wrappedModel, token);
+        }
+
+        /// <summary>
+        /// Retrieves a list of invoice comments appropriate to the filter.
+        /// </summary>
+        /// <param name="query">The filter.</param>
+        /// <param name="token">The token.</param>
+        /// <returns>
+        /// A task that represents the asynchronous operation.
+        /// The task result contains the filtered list of invoice comment.
+        /// </returns>
+        /// <exception cref="ArgumentException">Thrown when the parameter check fails.</exception>
+        /// <exception cref="NotAuthorizedException">Thrown when not authorized to access this resource.</exception>
+        /// <exception cref="NotFoundException">Thrown when the resource url could not be found.</exception>
+        public async Task<Types.PagedList<InvoiceComment>> GetCommentListAsync(Query<InvoiceComment, InvoiceCommentFilter> query, CancellationToken token = default)
+        {
+            if (query == null || query.Filter.InvoiceId <= 0 || query.Filter.InvoiceId == null)
+            {
+                throw new ArgumentException("Filter or a value of the Filter is null or invalid", nameof(query));
+            }
+            var jsonModel = await GetListAsync<InvoiceCommentListWrapper>("/api/invoice-comments", QueryString.For(query), token).ConfigureAwait(false);
+            return jsonModel.ToDomain();
+        }
+
+        /// <summary>
+        /// Retrieves an invoice comment by it's ID.
+        /// </summary>
+        /// <param name="id">The ID.</param>
+        /// <param name="token">The token.</param>
+        /// <returns>
+        /// A task that represents the asynchronous operation.
+        /// The task result contains the invoice comment.
+        /// </returns>
+        /// <exception cref="ArgumentException">Thrown when the parameter check fails.</exception>
+        /// <exception cref="NotAuthorizedException">Thrown when not authorized to access this resource.</exception>
+        /// <exception cref="NotFoundException">Thrown when the resource url could not be found.</exception>
+        public async Task<InvoiceComment> GetCommentByIdAsync(int id, CancellationToken token = default)
+        {
+            if (id <= 0)
+            {
+                throw new ArgumentException("invalid invoice comment id", nameof(id));
+            }
+            var jsonModel = await GetItemByIdAsync<InvoiceCommentWrapper>($"/api/invoice-comments/{id}", token).ConfigureAwait(false);
+            return jsonModel.ToDomain();
         }
     }
 }